New code updated

+++ app/src/services/profile/profile.context.js
@@ -8,113 +8,225 @@ import useHttp from '../../hooks/use-http';
 import {AuthenticationContext} from '../authentication/authentication.context';
 import {notificationActions} from '../../store/notification-slice';
 import {Alert} from 'react-native';
-import {setChangesMade} from '../../store/service-slice';
 import storage from '@react-native-firebase/storage';
+import {SQLiteContext} from '../sqlite/sqlite.context';
+import database from '@react-native-firebase/database';
+import {useNetInfo} from '@react-native-community/netinfo';
+import {loaderActions} from '../../store/loader-slice';
+import {getFirebaseAccessUrl} from '../../components/utility/helper';
+import RNFetchBlob from 'rn-fetch-blob';
+import {CameraRoll} from '@react-native-camera-roll/camera-roll';
 
 export const ProfileContext = createContext({
   onUpdateProfile: (data, successCallBack, errorCallback) => null,
   onRemoveProfilePicture: (successCallBack, errorCallback) => null,
   onUpdateProfilePicture: (successCallBack, errorCallback) => null,
+  onDownloadProfilePicture: (successCallBack, errorCallback) => null,
 });
 
 export const ProfileContextProvider = ({children}) => {
   const BACKEND_URL = remoteConfig().getValue('BACKEND_URL').asString();
   const dispatch = useDispatch();
   const {sendRequest} = useHttp();
+  const {updateData} = useContext(SQLiteContext);
+  const {isConnected} = useNetInfo();
 
-  const {
-    onLogout,
-    onGetUserDetails,
-    userData,
-    setUserData,
-    onSetUserAdditionalDetails,
-  } = useContext(AuthenticationContext);
+  const {onLogout, onGetUserDetails, userData} = useContext(
+    AuthenticationContext,
+  );
+
+  const showLoader = (loaderType, backdrop = true) => {
+    let options = {};
+    if (loaderType) {
+      options.loaderType = loaderType;
+    }
+    if (backdrop) {
+      options.backdrop = backdrop;
+    }
+    dispatch(loaderActions.showLoader(options));
+  };
+
+  const hideLoader = () => {
+    dispatch(loaderActions.hideLoader());
+  };
+
+  const showNotification = (status = 'error', message) => {
+    dispatch(
+      notificationActions.showToast({
+        status: status,
+        message: message,
+      }),
+    );
+  };
 
   const onUpdateProfile = async (
     data,
     successCallBack = () => {},
     errorCallback = () => {},
   ) => {
-    let currentUser = await auth().currentUser;
-    let providerId = currentUser.providerData[0].providerId;
-    let updatedDetails = {};
-    let {displayName, email} = data;
-
-    currentUser
-      .updateEmail(email)
-      .then(() => {
+    try {
+      if (!isConnected) {
+        throw 'No Internet Connection';
+      }
+      let currentUser = await auth().currentUser;
+      let {displayName, email} = data;
+      let updatedDetails = {};
+      if (
+        email === currentUser.email &&
+        displayName === currentUser.displayName
+      ) {
+        successCallBack();
+        return dispatch(
+          notificationActions.showToast({
+            status: 'warning',
+            message: 'No changes made to update',
+          }),
+        );
+      }
+      if (email !== currentUser.email) {
+        await currentUser.updateEmail(email);
         updatedDetails.email = email;
-        currentUser
-          .updateProfile({
-            displayName: displayName,
-          })
-          .then(() => {
-            updatedDetails.displayName = displayName;
-            onSuccessUpdatingProfileData(
-              updatedDetails,
-              successCallBack,
-              errorCallback,
-            );
-          })
-          .catch(err => {
-            errorCallback();
-            console.log('Error occured in updating profile data ' + err);
-            dispatch(
-              notificationActions.showToast({
-                message: 'Error in updating profile',
-                status: 'error',
-              }),
-            );
-          });
-      })
-      .catch(err => {
-        let error = '';
-        console.log(err.code, 'man error');
-        const showError = () => {
-          dispatch(
-            notificationActions.showToast({
-              message: error,
-              status: 'error',
-            }),
+      }
+
+      if (displayName !== currentUser.displayName) {
+        await currentUser.updateProfile({
+          displayName: displayName,
+        });
+        updatedDetails.displayName = displayName;
+      }
+
+      onSuccessUpdatingProfileData(
+        updatedDetails,
+        successCallBack,
+        errorCallback,
+      );
+      successCallBack(updatedDetails);
+      showNotification('success', 'Profile Updated');
+    } catch (err) {
+      let error = err.toString();
+      console.log(`Error occurred in updating profile: ${err.code}`, err);
+
+      const showError = () => {
+        dispatch(
+          notificationActions.showToast({
+            message: error,
+            status: 'error',
+          }),
+        );
+      };
+
+      switch (err.code) {
+        case 'auth/invalid-email':
+          error = 'Invalid email address';
+          showError();
+          break;
+        case 'auth/email-already-in-use':
+          error = 'Email-address is already in use! Try another email.';
+          showError();
+          break;
+        case 'auth/requires-recent-login':
+          Alert.alert(
+            'We have to identify its you?',
+            `For the security reasons, we should identify its you, so please re-login into our app again`,
+            [
+              {
+                text: 'RE-LOGIN',
+                onPress: () => {
+                  onLogout();
+                },
+                style: 'default',
+              },
+              {
+                text: 'CANCEL',
+                onPress: () => {},
+                style: 'cancel',
+              },
+            ],
           );
-        };
-        switch (err.code) {
-          case 'auth/invalid-email':
-            error = 'Invalid email address';
-            showError();
-            break;
-          case 'auth/email-already-in-use':
-            error = 'Email-address is already in use! Try another email.';
-            showError();
-            break;
-          case 'auth/requires-recent-login':
-            Alert.alert(
-              'We have to identify its you?',
-              `For the security reasons, we should identify its you, so please re-login into our app again`,
-              [
-                {
-                  text: 'RE-LOGIN',
-                  onPress: () => {
-                    onLogout();
-                  },
-                  style: 'default',
-                },
-                {
-                  text: 'CANCEL',
-                  onPress: () => {},
-                  style: 'cancel',
-                },
-              ],
-            );
-            break;
-          default:
-            error = 'Error Occured while updating email-address.';
-            showError();
-            break;
-        }
-        errorCallback();
-        console.log('Error occured in updating profile data email' + error);
-      });
+          break;
+        default:
+          error = 'Error Occured while updating ' + err.toString();
+          showError();
+          break;
+      }
+      errorCallback();
+    }
+
+    // currentUser
+    //   .updateEmail(email)
+    //   .then(() => {
+    //     updatedDetails.email = email;
+    //     currentUser
+    //       .updateProfile({
+    //         displayName: displayName,
+    //       })
+    //       .then(() => {
+    //         updatedDetails.displayName = displayName;
+    //         onSuccessUpdatingProfileData(
+    //           updatedDetails,
+    //           successCallBack,
+    //           errorCallback,
+    //         );
+    //       })
+    //       .catch(err => {
+    //         errorCallback();
+    //         console.log('Error occured in updating profile data ' + err);
+    //         dispatch(
+    //           notificationActions.showToast({
+    //             message: 'Error in updating profile',
+    //             status: 'error',
+    //           }),
+    //         );
+    //       });
+    //   })
+    //   .catch(err => {
+    //     let error = '';
+    //     console.log(err.code, 'man error');
+    //     const showError = () => {
+    //       dispatch(
+    //         notificationActions.showToast({
+    //           message: error,
+    //           status: 'error',
+    //         }),
+    //       );
+    //     };
+    //     switch (err.code) {
+    //       case 'auth/invalid-email':
+    //         error = 'Invalid email address';
+    //         showError();
+    //         break;
+    //       case 'auth/email-already-in-use':
+    //         error = 'Email-address is already in use! Try another email.';
+    //         showError();
+    //         break;
+    //       case 'auth/requires-recent-login':
+    //         Alert.alert(
+    //           'We have to identify its you?',
+    //           `For the security reasons, we should identify its you, so please re-login into our app again`,
+    //           [
+    //             {
+    //               text: 'RE-LOGIN',
+    //               onPress: () => {
+    //                 onLogout();
+    //               },
+    //               style: 'default',
+    //             },
+    //             {
+    //               text: 'CANCEL',
+    //               onPress: () => {},
+    //               style: 'cancel',
+    //             },
+    //           ],
+    //         );
+    //         break;
+    //       default:
+    //         error = 'Error Occured while updating email-address.';
+    //         showError();
+    //         break;
+    //     }
+    //     errorCallback();
+    //     console.log('Error occured in updating profile data email' + error);
+    //   });
   };
 
   //   calling this after updating the profile data in firebase to update details in backend
@@ -123,87 +235,52 @@ export const ProfileContextProvider = ({children}) => {
     successCallBack = () => {},
     errorCallback = () => {},
   ) => {
-    let currentUser = await auth().currentUser;
-
-    let jwtToken = await auth().currentUser.getIdToken();
-    sendRequest(
-      {
-        type: 'POST',
-        url: BACKEND_URL + '/user/',
-        data: {
-          ...details,
-        },
-        headers: {
-          authorization: 'Bearer ' + jwtToken,
-        },
-      },
-      {
-        successCallback: data => {
-          dispatch(setChangesMade({status: true}));
-          onSetUserAdditionalDetails(data.user);
-          setUserData(data.user);
-          currentUser.reload();
-          successCallBack();
-          dispatch(
-            notificationActions.showToast({
-              status: 'success',
-              message: 'Updated Profile Successfully',
-            }),
-          );
-        },
-        errorCallback: error => {
-          errorCallback();
-          dispatch(
-            notificationActions.showToast({
-              status: 'error',
-              message: error,
-            }),
-          );
-        },
-      },
-    );
+    try {
+      let currentUser = await auth().currentUser;
+      const uid = currentUser.uid;
+      let results = await updateData('Users', details, `WHERE uid=?`, [uid]);
+      await database().ref(`/users/${uid}`).update(details);
+      await onGetUserDetails();
+      currentUser.reload();
+      successCallBack();
+    } catch (err) {
+      errorCallback();
+    }
   };
 
   const onRemoveProfilePicture = async (
     successCallBack = () => {},
     errorCallback = () => {},
   ) => {
-    let jwtToken = await auth().currentUser.getIdToken();
-    sendRequest(
-      {
-        type: 'DELETE',
-        url: BACKEND_URL + '/user/remove-profile-picture',
-        headers: {
-          authorization: 'Bearer ' + jwtToken,
-        },
-      },
-      {
-        successCallback: async result => {
-          const onSuccess = () => {
-            dispatch(setChangesMade({status: true}));
-            successCallBack();
-            dispatch(
-              notificationActions.showToast({
-                message: result.message,
-                status: 'success',
-              }),
-            );
-          };
-          // call onsucces irrespective of geting user details failed or succefull
-          onGetUserDetails(onSuccess, onSuccess);
-        },
-        errorCallback: err => {
-          errorCallback();
-          console.log('Error in removing profile picture', err);
-          dispatch(
-            notificationActions.showToast({
-              message: err,
-              status: 'error',
-            }),
-          );
-        },
-      },
-    );
+    try {
+      if (!isConnected) {
+        throw 'No Internet Connection';
+      }
+      let photoURL = userData.photoURL;
+      let currentImageRef = storage().ref(photoURL);
+      let currentImageExists = await currentImageRef
+        .getMetadata()
+        .then(() => true)
+        .catch(() => false);
+      if (currentImageExists) {
+        await currentImageRef.delete();
+      }
+      await auth().currentUser.updateProfile({
+        photoURL: null,
+      });
+
+      onSuccessUpdatingProfileData(
+        {photoURL: null},
+        successCallBack,
+        errorCallback,
+      );
+      successCallBack();
+      showNotification('success', 'Profile picture removed successfully');
+    } catch (e) {
+      errorCallback();
+      hideLoader();
+      showNotification('error', e.toString());
+    }
   };
 
   const onUpdateProfilePicture = async (
@@ -211,17 +288,23 @@ export const ProfileContextProvider = ({children}) => {
     successCallBack = () => {},
     errorCallback = () => {},
   ) => {
-    let {extension, uri} = photo;
     try {
+      if (!isConnected) {
+        throw 'No Internet Connection';
+      }
+      let {extension, uri} = photo;
+
       // delete if image already exists
       let currentImagePath = userData.photoURL;
-      let currentImageRef = storage().ref(currentImagePath);
-      let currentImageExists = await currentImageRef
-        .getMetadata()
-        .then(() => true)
-        .catch(() => false);
-      if (currentImageExists) {
-        await currentImageRef.delete();
+      if (currentImagePath) {
+        let currentImageRef = storage().ref(currentImagePath);
+        let currentImageExists = await currentImageRef
+          .getMetadata()
+          .then(() => true)
+          .catch(() => false);
+        if (currentImageExists) {
+          await currentImageRef.delete();
+        }
       }
 
       let pictureName = `profile.${extension}`;
@@ -240,18 +323,55 @@ export const ProfileContextProvider = ({children}) => {
           successCallBack,
           errorCallback,
         );
+        showNotification('success', 'Profile picture updated successfully');
       } else {
         throw 'Error occured while uploading profile picture';
       }
     } catch (e) {
       errorCallback();
-      console.log('Error in updating profile picture', e);
-      dispatch(
-        notificationActions.showToast({
-          message: e.toString(),
-          status: 'error',
-        }),
+      console.error('Error in updating profile picture', e);
+      showNotification('error', e.toString());
+    }
+  };
+
+  const onDownloadProfilePicture = async (
+    successCallBack = () => {},
+    errorCallback = () => {},
+  ) => {
+    try {
+      if (!userData || !userData.photoURL) {
+        throw 'No Image Found';
+      }
+      if (!isConnected) {
+        throw 'No Internet connection';
+      }
+      // Getting the extention of the file
+      // get bytes
+      let photoURL = userData.photoURL;
+      let extension;
+      if (photoURL && photoURL.startsWith('users/')) {
+        photoURL = getFirebaseAccessUrl(userData.photoURL);
+        let extRegex = /\.(png|jpe?g|gif|bmp|webp)/i;
+        extension = photoURL.match(extRegex)?.[0];
+      }
+
+      showLoader('image_upload');
+
+      const res = await RNFetchBlob.config({
+        fileCache: true,
+        appendExt: extension,
+      }).fetch('GET', photoURL);
+      await CameraRoll.saveToCameraRoll(res.data);
+      showNotification(
+        'success',
+        'Profile Picture saved to your Gallery/Photos',
       );
+      hideLoader();
+      successCallBack();
+    } catch (e) {
+      errorCallback();
+      hideLoader();
+      showNotification('error', e.toString());
     }
   };
 
@@ -261,6 +381,7 @@ export const ProfileContextProvider = ({children}) => {
         onUpdateProfile,
         onRemoveProfilePicture,
         onUpdateProfilePicture,
+        onDownloadProfilePicture,
       }}>
       {children}
     </ProfileContext.Provider>
